+++ policykit/integrations/metagov/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from integrations.metagov import views
+
+
+urlpatterns = [
+    path('outcome/<int:id>', views.post_outcome),
+    path('action', views.action),
+]
